finish up impl

+++ src/otaclient/configs/_cfg_configurable.py
@@ -87,9 +87,9 @@ class _OTAClientSettings(BaseModel):
     READ_CHUNK_SIZE: int = 4 * 1024 * 1024  # 4MiB
 
     FSTRIM_AT_OTA: bool = True
-    FSTRIM_AT_OTA_TIMEOUT: float = 30  # seconds
+    FSTRIM_AT_OTA_TIMEOUT: int = 30  # seconds
     FSTRIM_AT_OTACLIENT_STARTUP: bool = True
-    FSTRIM_AT_OTACLIENT_STARTUP_TIMEOUT: float = 360  # 6mins
+    FSTRIM_AT_OTACLIENT_STARTUP_TIMEOUT: int = 360  # 6mins
 
 
 class _MultipleECUSettings(BaseModel):
